+++ src/main/java/ch/njol/skript/Skript.java
@@ -68,6 +68,9 @@ import org.skriptlang.skript.bukkit.log.runtime.BukkitRuntimeErrorConsumer;
 import org.skriptlang.skript.bukkit.registration.BukkitSyntaxInfos;
 import org.skriptlang.skript.bukkit.text.TextComponentParser;
 import org.skriptlang.skript.common.CommonModule;
+import org.skriptlang.skript.common.elements.functions.MathFunctions;
+import org.skriptlang.skript.common.elements.functions.StringFunctions;
+import org.skriptlang.skript.common.elements.functions.TimeFunctions;
 import org.skriptlang.skript.docs.Origin;
 import org.skriptlang.skript.lang.comparator.Comparator;
 import org.skriptlang.skript.lang.comparator.Comparators;
@@ -554,7 +557,12 @@ public final class Skript extends JavaPlugin implements Listener {
 					throw new RuntimeException(e);
 				}
 				BukkitEventValues.register(eventValueRegistry);
-				new DefaultFunctions();
+				{
+					// load functions
+					new MathFunctions();
+					new StringFunctions();
+					new TimeFunctions();
+				}
 				new DefaultOperations();
 			}
