+++ src/main/java/org/skriptlang/skript/bukkit/BukkitModule.java
@@ -11,6 +11,7 @@ import org.skriptlang.skript.bukkit.brewing.BrewingModule;
 import org.skriptlang.skript.bukkit.damagesource.DamageSourceModule;
 import org.skriptlang.skript.bukkit.entity.EntityModule;
 import org.skriptlang.skript.bukkit.fishing.FishingModule;
+import org.skriptlang.skript.bukkit.functions.BukkitFunctionModule;
 import org.skriptlang.skript.bukkit.input.InputModule;
 import org.skriptlang.skript.bukkit.item.ItemModule;
 import org.skriptlang.skript.bukkit.itemcomponents.ItemComponentModule;
@@ -38,6 +39,7 @@ public class BukkitModule extends HierarchicalAddonModule {
 			new BlockModule(this),
 			new BreedingModule(this),
 			new BrewingModule(this),
+			new BukkitFunctionModule(this),
 			new DamageSourceModule(this),
 			new EntityModule(this),
 			new FishingModule(this),
